@@ -1,20 +1,21 @@
 {
   "name": "MK-Web",
   "version": "1.0.0",
+  "type": "module",
   "private": true,
   "scripts": {
     "dev": "next dev --port 3000",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "eslint .",
     "format": "prettier --write ."
   },
   "dependencies": {
     "@hookform/resolvers": "^3.9.0",
     "@mdx-js/loader": "^3.1.1",
     "@mdx-js/react": "^3.1.1",
-    "@next/mdx": "^16.0.3",
-    "@next/third-parties": "^15.4.0-canary.51",
+    "@next/mdx": "16.0.10",
+    "@next/third-parties": "16.0.10",
     "@radix-ui/react-dropdown-menu": "^2.1.2",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-label": "^2.1.0",
@@ -26,8 +27,8 @@
     "@types/mdx": "^2.0.13",
     "@types/node": "20.6.2",
     "@types/nodemailer": "^6.4.16",
-    "@types/react": "^18.3.11",
-    "@types/react-dom": "^18.3.0",
+    "@types/react": "19.2.7",
+    "@types/react-dom": "19.2.3",
     "@upstash/ratelimit": "^0.4.4",
     "@vercel/analytics": "^1.3.1",
     "@vercel/blob": "^0.12.5",
@@ -41,15 +42,15 @@
     "gray-matter": "^4.0.3",
     "lucide-react": "^0.451.0",
     "nanoid": "^5.0.7",
-    "next": "^15.3.2",
+    "next": "16.0.10",
     "next-plausible": "^3.12.2",
     "next-recaptcha-v3": "^1.4.1",
     "next-themes": "^0.3.0",
     "nodemailer": "^6.9.15",
     "postcss": "8.4.29",
     "prettier": "^3.3.3",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "19.2.3",
+    "react-dom": "19.2.3",
     "react-hook-form": "^7.53.0",
     "react-hot-toast": "^2.4.1",
     "remark-frontmatter": "^5.0.0",
@@ -63,8 +64,12 @@
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.15",
-    "eslint": "^8.49.0",
-    "eslint-config-next": "^14.2.14"
+    "eslint": "^9",
+    "eslint-config-next": "16.0.10"
   },
-  "packageManager": "pnpm@9.12.2+sha512.22721b3a11f81661ae1ec68ce1a7b879425a1ca5b991c975b074ac220b187ce56c708fe5db69f4c962c989452eee76c82877f4ee80f474cebd61ee13461b6228"
+  "packageManager": "pnpm@9.12.2+sha512.22721b3a11f81661ae1ec68ce1a7b879425a1ca5b991c975b074ac220b187ce56c708fe5db69f4c962c989452eee76c82877f4ee80f474cebd61ee13461b6228",
+  "overrides": {
+    "@types/react": "19.2.7",
+    "@types/react-dom": "19.2.3"
+  }
 }
